Move generic_sesh to tests/helpers.py
