+++ tests/weather_anomaly/details/test_monthly_views_common.py
@@ -29,8 +29,7 @@ import datetime
 
 from pytest import fixture, mark, approx
 
-from ...helpers import create_then_drop_views
-from pycds.util import generic_sesh
+from ...helpers import generic_sesh, create_then_drop_views
 from pycds import Obs
 from pycds.weather_anomaly import \
     MonthlyAverageOfDailyMaxTemperature, MonthlyAverageOfDailyMinTemperature, \
